requirements.txt
# Quant Agent v2.1 -- Public Edition

# AI
anthropic>=0.34.0

# Market Data
yfinance>=0.2.31
pandas>=2.1.0
numpy>=1.25.0

# API Server (Dashboard)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
